#430: the unit-value domain gate negated the domain instead of comparing it
THE DEFECT
AI_unitValue's gate read `!kUnitInfo.getDomain() == AI_unitAIDomainType(eUnitAI)`,
so the `!` bound to the domain rather than to the comparison -- a bool weighed
against a DomainTypes. With SEA=0/AIR=1/LAND=2/IMMOBILE=3 it happened to reject
non-sea units for sea roles and sea units for air roles, while letting LAND units
be valued for AIR roles, AIR for LAND and SEA for LAND, every one of which should
have returned 0. It sits at the head of the AI's unit valuation, so the mis-gate
reached every build decision. A behaviour change, stated rather than tuned around.

THE EMPTY BODY
processUnitCombat carried `if (bAdding && iClassGroupBase > -10) { }` -- the shell
of the per-class group-rank accumulation the info's load-derived sums replaced. The
comment above it says the ranks ARE those sums and must never be re-derived per
class, and the const had no other reader, so this was a deliberate removal's
residue rather than a lost body. Same class as the bodyless for-loop in 3c850baa.

THE EXE BOUNDARY
CvCity::getBuildingVisibilityPriority is DllExport and EXE-resolved, so its float
return is a fixed ABI obligation while the info holds the priority as an int. The
conversion is genuine and is now explicit.

THE FROZEN HEADER
Boost 1.32 declares dict_base dll-interface over a non-dll-interface api::object.
python::dict is used in both including files so the header cannot be dropped, and
the toolchain is frozen by the closed EXE so it cannot be fixed. Guarded at the
include alone, leaving C4275 live for our own declarations -- the same disposition
CvString.h already takes for its sibling C4251.

DEAD LOCALS
Three BTS-era top-of-function declarations in CvUnitAI with no remaining reader.

VERIFIED: Release rebuild, 0 warnings and 0 errors, from 9 warnings across 4 codes.
Line endings re-checked on every touched file: pure CRLF, zero bare LF.

+++ Sources/Engine/CvCity.cpp
@@ -3901,7 +3901,9 @@ const CvArtInfoBuilding* CvCity::getBuildingArtInfo(BuildingTypes eBuilding) con
 
 float CvCity::getBuildingVisibilityPriority(BuildingTypes eBuilding) const
 {
-	return GC.getBuildingInfo(eBuilding).getVisibilityPriority();
+	// The float is the EXE's: this accessor is DllExport and the closed EXE resolves it, so the return type is a
+	// fixed ABI obligation while the info holds the priority as an int. The conversion is the boundary, made explicit.
+	return static_cast<float>(GC.getBuildingInfo(eBuilding).getVisibilityPriority());
 }
 
 bool CvCity::hasTrait(TraitTypes eTrait) const
